refactor: date change widget
