+++ lib/component/provider.dart
@@ -1,10 +1,10 @@
-// Copyright 2023 Fries_I23
+// Copyright 2024 Fries_I23
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//     https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,12 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export './appbar_combined_action.dart'
-    show AppbarActionShowStatus, AppBarActions, AppbarActionItemConfig;
-export './colorful_navibar.dart' show ColorfulNavibar;
-export './contributor_tile.dart';
-export 'date_changer.dart';
-export './donate_dialog.dart' show DonateDialogResult;
-export './fixed_page_place_holder.dart';
-export './habit_divider.dart';
-export './not_found_image.dart';
+export './widgets/date_change.dart' show DateChangeNotifier;
