+++ lib/view/page_habits_display.dart
@@ -44,7 +44,6 @@ import '../persistent/local/handler/habit.dart';
 import '../provider/app_compact_ui_switcher.dart';
 import '../provider/app_developer.dart';
 import '../provider/app_theme.dart';
-import '../provider/habit_date_change.dart';
 import '../provider/habit_op_config.dart';
 import '../provider/habit_summary.dart';
 import '../provider/habits_file_exporter.dart';
@@ -439,7 +438,7 @@ class _HabitsDisplayView extends State<HabitsDisplayView>
 
   Future<void> _onRefreshIndicatorTriggered() async {
     if (!mounted) return;
-    HabitDateChangeProvider.of(context).dateTime = HabitDate.now();
+    DateChangeProvider.of(context).dateTime = HabitDate.now();
     context.read<HabitSummaryViewModel>().rockreloadDBToggleSwich();
   }
 
@@ -738,7 +737,7 @@ class _HabitsDisplayView extends State<HabitsDisplayView>
         viewmodel.isCalendarExpanded,
         viewmodel.getHabitInsideVersion(uuid),
         viewmodel.isHabitSelected(uuid),
-        HabitDateChangeProvider.of(context).dateTime,
+        DateChangeProvider.of(context).dateTime,
       ),
       shouldRebuild: (previous, next) => previous != next,
       builder: (context, contents, child) => Selector<AppThemeViewModel, int>(
@@ -816,7 +815,7 @@ class _HabitsDisplayView extends State<HabitsDisplayView>
                 verticalScrollController: viewmodel.verticalScrollController,
                 horizonalScrollControllerGroup:
                     viewmodel.horizonalScrollControllerGroup,
-                startDate: HabitDateChangeProvider.of(context).dateTime,
+                startDate: DateChangeProvider.of(context).dateTime,
                 endDate: viewmodel.earliestSummaryDataStartDate?.startDate,
                 isExtended: isExtended,
                 collapsePrt: occupyPrt,
@@ -910,7 +909,7 @@ class _HabitsDisplayView extends State<HabitsDisplayView>
                 verticalScrollController: viewmodel.verticalScrollController,
                 horizonalScrollControllerGroup:
                     viewmodel.horizonalScrollControllerGroup,
-                startDate: HabitDateChangeProvider.of(context).dateTime,
+                startDate: DateChangeProvider.of(context).dateTime,
                 endDate: viewmodel.earliestSummaryDataStartDate?.startDate,
                 isExtended: isExtended,
                 collapsePrt: occupyPrt,
